backend/requirements.txt
flask==3.0.3
langchain==0.3.11
langchain-community==0.3.11
langchain-text-splitters==0.3.2
langchain-google-vertexai==2.0.4
google-cloud-aiplatform>=1.111.0,<2.0.0
pypdf==5.0.1
numpy==1.26.4
python-dotenv==1.0.1

